Conflicts:
	arch/arm/configs/msm8960_mmi_defconfig
	drivers/char/random.c
	drivers/gpu/drm/i915/intel_display.c
	drivers/media/video/msm/msm.c
	fs/nfs/nfs4proc.c
	kernel/cgroup.c
	scripts/kconfig/Makefile
	scripts/kconfig/conf.c

Merge razrqcom-dev-team rebase-y branch

+++ fs/ecryptfs/miscdev.c
@@ -431,9 +431,9 @@ ecryptfs_miscdev_write(struct file *file, const char __user *buf,
 		   || count > (1 + 4 + 2 + sizeof(struct ecryptfs_message) + 4
 			       + ECRYPTFS_MAX_ENCRYPTED_KEY_BYTES)) {
 		printk(KERN_WARNING "%s: Acceptable packet size range is "
-		       "[%d-%u], but amount of data written is [%zu].",
+		       "[%d-%lu], but amount of data written is [%zu].",
 		       __func__, (1 + 4 + 1),
-		       (1 + 4 + 2 + sizeof(struct ecryptfs_message) + 4
+		       (long unsigned int)(1 + 4 + 2 + sizeof(struct ecryptfs_message) + 4
 			+ ECRYPTFS_MAX_ENCRYPTED_KEY_BYTES), count);
 		return -EINVAL;
 	}
